Refactor: js code 리팩토링
