+++ week2_assignment/shopping/css/week1_css/index.css
@@ -153,7 +153,7 @@ main {
     scroll-margin-left: 14.375rem;
 }
 
-.container {
+.shoppingList_container {
     display: grid;
     gap: 1rem 1rem;
     width: 83.4%;
@@ -206,24 +206,24 @@ main {
 }
 
 @media screen and (min-width: 1320px) {
-    .container {
+    .shoppingList_container {
         grid-template-columns: repeat(5, 12.5rem);
     }
 }
 @media screen and (max-width: 1320px) {
-    .container {
+    .shoppingList_container {
         grid-template-columns: repeat(4, 12.5rem);
     }
 }
 
 @media screen and (max-width: 1100px) {
-    .container {
+    .shoppingList_container {
         grid-template-columns: repeat(3, 12.5rem);
     }
 }
 
 @media screen and (max-width: 880px) {
-    .container {
+    .shoppingList_container {
         grid-template-columns: repeat(2, 17.5rem);
     }
     .item {
@@ -239,7 +239,7 @@ main {
 }
 
 @media screen and (max-width: 810px) {
-    .container {
+    .shoppingList_container {
         grid-template-columns: repeat(1, 31.25rem);
     }
