+++ week2_assignment/shopping/css/week2_css/cart.css
@@ -131,17 +131,19 @@ tbody td:nth-child(n + 2) {
 .home_btn {
     background-color: #c77dff;
     color: whitesmoke;
+    text-align: center;
+    line-height: 2rem;
 }
 
 /** modal css **/
 .modal_background {
     display: none;
     position: fixed;
-    z-index: 1000;
     top: 0;
-    left: 0;
-    bottom: 0;
     right: 0;
+    bottom: 0;
+    left: 0;
+    z-index: 1000;
     background: rgba(0, 0, 0, 0.5);
 }
