+++ week2_assignment/shopping/css/week2_css/index.css
@@ -28,5 +28,3 @@
         background-color: white;
     }
 }
-
-/** left_nav css **/
