+++ week2_assignment/shopping/html/cart.html
@@ -5,6 +5,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link rel="stylesheet" href="../css/reset.css" />
         <link rel="stylesheet" href="../css/week2_css/cart.css" />
+        <link rel="stylesheet" href="../css/week2_css/index.css" />
         <title>CartList</title>
     </head>
     <body>
@@ -17,6 +18,26 @@
                 <img src="../img/hambuger-icon.png" alt="햄버거 아이콘" />
             </button>
         </header>
+        <section class="menu_bar">
+            <nav class="menu_nav">
+                <div>
+                    <button type="button" class="menu_hide_btn btn">
+                        <i class="fa-solid fa-right-long fa-2x"></i>
+                    </button>
+                </div>
+                <ul>
+                    <li>
+                        <a>관심 상품 목록</a>
+                    </li>
+                    <li>
+                        <a>관리자 페이지</a>
+                    </li>
+                    <li>
+                        <a href="./cart.html">장바구니</a>
+                    </li>
+                </ul>
+            </nav>
+        </section>
         <main>
             <table id="cart_table">
                 <thead>
@@ -48,7 +69,7 @@
 
             <section class="btn_section">
                 <button class="buy_btn cart_btn">구매하기</button>
-                <button class="home_btn cart_btn">홈으로</button>
+                <a class="home_btn cart_btn" href="./index.html">홈으로</a>
             </section>
 
             <section class="modal_background">
@@ -62,6 +83,12 @@
                 </div>
             </section>
         </main>
+        <script
+            src="https://kit.fontawesome.com/842f270e6e.js"
+            crossorigin="anonymous"
+        ></script>
+        <script src="../js/menuBar.js" type="module"></script>
+        <script src="../js/modal.js" type="module"></script>
         <script src="../js/cart.js" type="module"></script>
     </body>
 </html>
