+++ week2_assignment/shopping/html/index.html
@@ -97,7 +97,7 @@
             </section>
 
             <section class="shoppingList">
-                <section class="container"></section>
+                <section class="shoppingList_container"></section>
             </section>
         </main>
 
@@ -111,7 +111,6 @@
 
         <script src="../js/menuBar.js" type="module"></script>
         <script src="../js/itemData.js" type="module"></script>
-        <script src="../js/itemFilter.js" type="module"></script>
         <script src="../js/index.js" type="module"></script>
     </body>
 </html>
