+++ week2_assignment/shopping/js/index.js
@@ -1,45 +1,38 @@
-// 상품 클릭시 장바구니에 추가해주는 함수
-export function addItem() {
-    if (confirm("장바구니에 추가하겠습니까?") === true) {
-        const itemTitle = this.children[2].innerText;
+import SHOPPING_LIST from "./itemData.js";
+import { addItemToCart } from "./addItemToCart.js";
+import { createShoppingList } from "./createElementList.js";
 
-        if (!localStorage.getItem("cartList")) {
-            localStorage.setItem("cartList", JSON.stringify([itemTitle]));
-        } else {
-            let cartList = JSON.parse(localStorage.getItem("cartList"));
+const section = document.querySelector(".shoppingList_container");
 
-            // 장바구니 상품 중복 방지
-            cartList = cartList.includes(itemTitle)
-                ? cartList
-                : [...cartList, itemTitle];
-            localStorage.setItem("cartList", JSON.stringify(cartList));
-        }
-        window.location.href = "../html/cart.html";
+// 왼쪽 nav바의 버튼들에 상품 필터링 이벤트 적용
+const btnList = document.querySelectorAll(".nav_btn");
+btnList.forEach((elem) => {
+    elem.addEventListener("click", () => {
+        showShoppingList(`${elem.className.split("_")[0]}`);
+    });
+});
+
+// 상품 필터링해주고 화면에 보여주는 함수
+function showShoppingList(category) {
+    section.innerHTML = "";
+    let filterdList;
+
+    // category에 따른 상품 필터링
+    if (category === "total") filterdList = SHOPPING_LIST;
+    else {
+        filterdList = SHOPPING_LIST.filter(
+            (elem) => elem.category === category
+        );
     }
+
+    // html에 shopping 리스트 생성
+    const list = createShoppingList(filterdList);
+    list.forEach((elem) => (section.innerHTML += elem));
+
+    // 삽입된 상품 요소에 click시 장바구니추가 이벤트 적용
+    const items = document.querySelectorAll(".item");
+    items.forEach((elem) => elem.addEventListener("click", addItemToCart));
 }
 
-// HTML파일에 삽입될 상품 element 배열 생성 함수
-export function createShoppingList(filterdList) {
-    return filterdList.map((elem) => {
-        return `
-            <div class="item ${elem.category}">
-                <img src=${elem.src} alt=${elem.title}/>
-                <span><i class="fa-regular fa-heart love"></i></span>
-                <span>${elem.title}</span>
-                <span>${elem.price.toLocaleString()}</span>
-            </div>
-        `;
-    });
-}
-
-export function createPurchaseList(purchaseList) {
-    return purchaseList.map((elem) => {
-        return `
-            <div class="purchase_item">
-                <img src=${elem.src} alt=${elem.title} />
-                <p>${elem.title}</p>
-                <p>${elem.price.toLocaleString()}</p>
-            </div>
-        `;
-    });
-}
+// 첫 화면에서 전체 상품 랜더링
+showShoppingList("total");
